Apply recursive deep_fetch
